Refactored stream operations into distinct headers.
