+++ src/manip/terminal_streamer.cpp
@@ -1,5 +1,5 @@
-#include "terminalpp/terminal_manip.hpp"
-#include "terminalpp/terminal.hpp"
+#include "terminalpp/manip/terminal_streamer.hpp"
+#include "terminalpp/manip/write_string.hpp"
 
 namespace terminalpp {
